requirements.txt
Flask
oracledb
python-dotenv
cryptography
